requirements.txt
# Snowflake
snowflake-connector-python==3.10.0

# Kafka (confluent-kafka est plus moderne et compatible Python 3.12)
confluent-kafka==2.3.0

# PostgreSQL (for testing)
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# Retry logic et resilience
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality (dev dependencies)
flake8==6.1.0
black==23.12.1
isort==5.13.2
mypy==1.7.1

